services/ingestion/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
watchdog==3.0.0
python-dotenv==1.0.0
pytesseract==0.3.10
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0
requests==2.31.0
